+++ src/types.ts
@@ -116,6 +116,7 @@ export const LANGUAGES = [
   'vbnet',
   'erlang',
   'terraform',
+  'crystal',
   'unknown',
 ] as const;